Add policy layer

+++ cmd/web/apps/youtube_ex_api/lib/youtube_ex_api/controllers/user_controller.ex
@@ -20,23 +20,33 @@ defmodule YoutubeExApi.UserController do
   end
 
   def show(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
-    render(conn, "show.json", user: user)
-  end
-
-  def update(conn, %{"id" => id} = user_params) do
-    user = Accounts.get_user!(id)
-
-    with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
+    with :ok <- Accounts.can_show_user?(conn.assigs.current_user.id) do
+      user = Accounts.get_user!(id)
       render(conn, "show.json", user: user)
     end
   end
 
-  def delete(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
+  def update(conn, %{"id" => id} = user_params) do
+    if Accounts.can_update_user?(conn.assigs.current_user.id) do
+      user = Accounts.get_user!(id)
 
-    with {:ok, %User{}} <- Accounts.delete_user(user) do
-      send_resp(conn, :no_content, "")
+      with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
+        render(conn, "show.json", user: user)
+      end
+    else
+      {:error, :forbidden}
+    end
+  end
+
+  def delete(conn, %{"id" => id}) do
+    if Accounts.can_update_user?(conn.assigs.current_user.id) do
+      user = Accounts.get_user!(id)
+
+      with {:ok, %User{}} <- Accounts.delete_user(user) do
+        send_resp(conn, :no_content, "")
+      end
+    else
+      {:error, :forbidden}
     end
   end
 end
